排版优化，增加对bootstrapnodes为空的判断

+++ libethereum/NodeConnParamsManagerSSL.cpp
@@ -317,28 +317,35 @@ void NodeConnParamsManagerSSL::updateBootstrapnodes()const
             } 
             root["nodes"] = nodes;
         }
-        else
+        else {
             root["nodes"].resize(0);
+        }
         
-	try {
-		string content = w.write(root);
-		writeFile(getDataDir() + "/bootstrapnodes.json.new",content,true);
-		if(contentsString(getDataDir() + "/bootstrapnodes.json.new") == content) {
-			boost::filesystem::path oldPath(getDataDir() + "/bootstrapnodes.json.new");
-			boost::filesystem::path newPath(getDataDir() + "/bootstrapnodes.json");
+		try {
+			string content = w.write(root);
 
-			boost::filesystem::remove(newPath);
-			boost::filesystem::rename(oldPath, newPath);
-		}
-		else {
-			LOG(ERROR) << "Write bootstrapnodes fail!";
-		}
+			if(content.empty()) {
+				LOG(ERROR) << "Empty bootstrapnodes content";
+				return;
+			}
 
-		LOG(INFO) << "updateBootstrapnodes:  " << content ;
-	}
-	catch(std::exception &e) {
-		LOG(ERROR) << "Write bootstrapnodes fail: " << e.what();
-	}
+			writeFile(getDataDir() + "/bootstrapnodes.json.new",content,true);
+			if(contentsString(getDataDir() + "/bootstrapnodes.json.new") == content) {
+				boost::filesystem::path oldPath(getDataDir() + "/bootstrapnodes.json.new");
+				boost::filesystem::path newPath(getDataDir() + "/bootstrapnodes.json");
+
+				boost::filesystem::remove(newPath);
+				boost::filesystem::rename(oldPath, newPath);
+			}
+			else {
+				LOG(ERROR) << "Write bootstrapnodes fail!";
+			}
+
+			LOG(INFO) << "updateBootstrapnodes:  " << content ;
+		}
+		catch(std::exception &e) {
+			LOG(ERROR) << "Write bootstrapnodes fail: " << e.what();
+		}
     }
     catch (std::exception &e)
     {
